refactor mongodb utils name
