+++ src/app/service/bounty/RefreshBounty.ts
@@ -8,7 +8,7 @@ import { addSubmitReactions } from './SubmitBounty';
 import { addCompletedReactions } from './CompleteBounty';
 import BountyUtils from '../../utils/BountyUtils';
 import Log from '../../utils/Log';
-import MongoDbUtils from '../../utils/dbUtils';
+import MongoDbUtils from '../../utils/MongoDbUtils';
 
 /**
  * This service will refresh the bounty in the Bounty board with the correct information
