+++ src/app/service/guest-pass/RemoveGuestPass.ts
@@ -3,7 +3,7 @@ import constants from '../constants/constants';
 import ServiceUtils from '../../utils/ServiceUtils';
 import { GuildMember } from 'discord.js';
 import Log, { LogUtils } from '../../utils/Log';
-import MongoDbUtils from '../../utils/dbUtils';
+import MongoDbUtils from '../../utils/MongoDbUtils';
 
 export default async (guestUser: GuildMember): Promise<any> => {
 	if (guestUser.user.bot) {
